src/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pbkdf2==1.3
pycryptodome==3.20.0
pycryptodomex==3.20.0
requests==2.31.0
urllib3==2.2.0
websockets==12.0
